Namespaces refactoring + migration + fixes
