+++ src/Resources/config/routing.yml
@@ -1,9 +1,9 @@
 sylius_refund_admin:
-    resource: "admin_routing.yml"
+    resource: "@SyliusRefundPlugin/Resources/config/admin_routing.yml"
     prefix: /admin
 
 sylius_refund_shop:
-    resource: "shop_routing.yml"
+    resource: "@SyliusRefundPlugin/Resources/config/shop_routing.yml"
     prefix: /{_locale}
-        requirements:
-            _locale: ^[a-z]{2}(?:_[A-Z]{2})?$
+    requirements:
+        _locale: ^[a-z]{2}(?:_[A-Z]{2})?$
